workfiles/requirements.txt
requests>=2.31.0
urllib3>=2.0.7
numpy>=1.24.0
openai-whisper>=20231117
python-dotenv>=1.0.0 